Established general architecture pattern, ported data into DB, will be branching commits from here on

+++ cruncher_api/src/repos/similarity_ranking_repo/similarity_ranking_models.rs
@@ -1,15 +1,17 @@
 use super::super::super::schema::similarity_ranking;
-#[derive(Queryable)]
+#[derive(Queryable, Identifiable, Debug, Clone)]
+#[primary_key(tickera, tickerb)]
+#[table_name = "similarity_ranking"]
 pub struct QueryableSimilarityRanking {
     pub tickera: String,
     pub tickerb: String,
-    pub similarity: Numeric,
+    pub similarity: i32,
 }
 
-#[derive(Insertable)]
+#[derive(Insertable, AsChangeset)]
 #[table_name = "similarity_ranking"]
 pub struct InsertableSimilarityRanking<'a> {
     pub tickera: &'a str,
     pub tickerb: &'a str,
-    pub similarity: &'a f32,
+    pub similarity: &'a i32,
 }
